replaced BigDecimals to float

+++ src/main/java/com/example/dto/order/OrderResponse.java
@@ -4,7 +4,6 @@ import com.example.dto.orderItem.OrderItemResponse;
 import com.example.models.OrderStatus;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,5 +15,5 @@ public class OrderResponse {
     private LocalDateTime orderDate;
     private OrderStatus status;
     private List<OrderItemResponse> items;
-    private BigDecimal totalAmount;
+    private float totalAmount;
 }
