+++ src/main/java/com/example/dto/orderItem/OrderItemResponse.java
@@ -1,7 +1,6 @@
 package com.example.dto.orderItem;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class OrderItemResponse {
@@ -9,6 +8,6 @@ public class OrderItemResponse {
     private Long productId;
     private String productName;
     private int quantity;
-    private BigDecimal unitPrice;
-    private BigDecimal totalPrice;
+    private float unitPrice;
+    private float totalPrice;
 }
